requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alibabacloud-credentials==0.3.2
alibabacloud-dysmsapi20170525==2.0.23
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.2.1
alibabacloud-tea==0.3.1
alibabacloud-tea-openapi==0.3.7
alibabacloud-tea-util==0.3.8
alibabacloud-tea-xml==0.0.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-redis==5.2.0
django-rest-swagger==2.2.0
django-tinymce==3.6.1
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.4
openapi-codec==1.3.2
packaging==23.1
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.3
pytz==2023.3
redis==4.5.4
requests==2.28.2
ruamel.yaml==0.17.21
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.8.2
